requirements.txt
# fastapi==0.118.0
# uvicorn[standard]==0.24.0
# langchain==0.3.27
# langchain_community==0.3.30
# langchain_core==0.3.78
# langchain_google_genai==2.1.12
# langchain_groq==0.3.8
# langchain_huggingface==0.3.1
# langchain_text_splitters==0.3.11
# langgraph==0.6.8
# pydantic==2.11.10
# python-dotenv==1.1.1
# supabase==2.21.1


fastapi==0.118.0
uvicorn[standard]==0.24.0
langchain==0.3.27
langchain_community==0.3.30
langchain_core==0.3.78
langchain_google_genai==2.1.12
langchain_groq==0.3.8
langchain_huggingface==0.3.1
langchain_text_splitters==0.3.11
langgraph==0.6.8
pydantic==2.11.10
python-dotenv==1.1.1
supabase

# ADD THESE MISSING PACKAGES:
sentence-transformers==2.6.0
pypdf
# transformers==4.37.2
# torch==2.1.2
# numpy==1.24.4
# faiss-cpu==1.7.4
